+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project
 adheres to [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+### :boom: Breaking changes
+
+- Update code with CPython 3.14.0 release candidate 1 version
+
 ## [0.3.0] - 2025-07-17
 
 [0.3.0]: https://github.com/rogdham/backports.zstd/releases/tag/v0.2.0